pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "edilcloud-back-dn"
version = "0.1.0"
description = "EdilCloud backend v3 bootstrap"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "Django>=5.2,<5.3",
  "channels>=4.1,<5",
  "channels-redis>=4.2,<5",
  "django-ninja>=1.3,<2",
  "django-cors-headers>=4.7,<5",
  "dj-database-url>=2.2,<3",
  "psycopg[binary]>=3.2,<4",
  "PyJWT>=2.10,<3",
  "redis[hiredis]>=6,<7",
  "httpx>=0.28,<0.29",
  "requests>=2.32,<3",
  "google-auth>=2.38,<3",
  "pgvector>=0.3,<0.4",
  "python-dotenv>=1.0,<2",
  "openpyxl>=3.1,<4",
  "xlrd>=2.0,<3",
  "jpype1>=1.6,<2",
  "mpxj>=16.1,<17",
  "gunicorn>=23,<24",
  "uvicorn>=0.34,<0.39",
  "websockets>=14,<16",
  "whitenoise>=6.9,<7",
  "stripe>=13.1,<14",
  "sentry-sdk>=2.27,<3",
  "opentelemetry-api>=1.33,<2",
  "opentelemetry-sdk>=1.33,<2",
  "opentelemetry-instrumentation-django>=0.54b0,<1",
  "opentelemetry-instrumentation-psycopg>=0.54b0,<1",
]

[project.optional-dependencies]
dev = [
  "django-debug-toolbar>=5.2,<6",
  "django-extensions>=4.1,<5",
  "pytest>=8.3,<9",
  "pytest-django>=4.11,<5",
  "ruff>=0.11,<0.12",
]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"edilcloud.modules.assistant" = ["data/*.json"]
"edilcloud.modules.identity" = ["templates/**/*.txt", "templates/**/*.html"]
"edilcloud.modules.workspaces" = ["templates/**/*.txt", "templates/**/*.html"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "edilcloud.settings.test"
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-p no:cacheprovider"

[tool.ruff]
line-length = 100
target-version = "py312"
